fix: close SQL injection across e107 core via input escaping
Backport of the core SQL-injection remediation to the 2.x line, which
predates the e_db prepared-statement API. The same sinks flagged by the
full-core audit are neutralised with legacy-compatible techniques.

Techniques:
- escape() applied to tainted values inside their SQL quotes.
- (int) casts for LIMIT/OFFSET and numeric columns.
- strict [A-Za-z0-9_] allowlists for dynamic table and column
  identifiers, which escaping cannot protect.
- Import providers fixed at the root in base_import_class::database()
  (backtick-doubling of the database name plus prefix sanitisation).
- alt_auth otherdb/e107db login fixed with PDO quote() plus identifier
  validation (advisory GHSA-5f75-c25x-59wx).

Sinks were located by content against the 2.x tree (the audit line
numbers come from master). DB abstraction-layer methods that rely on a
documented trusted-fragment contract are left unchanged.

Adds SqlInjectionFixesTest. Unit suite green.

Second audit pass (independent re-scan) added three sinks the first pass
did not flag, fixed with the same legacy-compatible approach:
- alt_auth admin config (alt_auth_adminmenu.php): the option-save loop
  interpolated attacker-controlled $_POST key NAMES into select/update/
  insert; values now passed through escape().
- menumanager menuDeactivate: menu_id taken from $_POST['removeid'] was
  not cast before reaching gen/delete/update; now (int).
- user mailout: the last_visit_date selector slipped past a loose numeric
  guard into an unquoted comparison; now (int).
All three require an authenticated admin (or CSRF against one).

+++ e107_plugins/forum/shortcodes/batch/forum_shortcodes.php
@@ -245,7 +245,7 @@ class forum_shortcodes extends e_shortcode
 
 		if(USER == true)
 		{
-			$total_new_threads = defined('USERLV') ? e107::getDb()->count('forum_thread', '(*)', "WHERE thread_datestamp>'" . USERLV . "' ") : 0;
+			$total_new_threads = defined('USERLV') ? e107::getDb()->count('forum_thread', '(*)', "WHERE thread_datestamp>'" . (int) USERLV . "' ") : 0;
 			$total_read_threads = 0;
 
 			if(defined('USERVIEWED') && defset('USERVIEWED') != "")
